backend/pyproject.toml
[project]
name = "prolix-backend"
version = "0.1.0"
description = "Prolix MVP backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.0",
  "psycopg[binary]>=3.2.0",
  "pgvector>=0.3.0",
  "pydantic>=2.7.0",
  "python-multipart>=0.0.9",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.2.0", "httpx>=0.27.0"]

[tool.pytest.ini_options]
pythonpath = ["app"]
